# cloudnet-rest-module/src/main/resources/documentation/swagger.yaml
openapi: 3.1.0
info:
  description: >
    CloudNet's rest API version 3 is the next generation in CloudNet's rest
    history.
    
    
    In order to access rest API v3 you will have to authenticate yourself using basic auth.
    To create a basic auth user use the `rest user create <id> <password>` command on your cloudnet node.
    After authenticating with basic auth on `/auth` you will get a bearer jwt access token that must be used for further
    requests to the api. For further information about the response take a look at the `/auth` endpoint.
    
    MENTION CONFIG HERE
    MENTION CASE-SENS HERE
  version: '3'
  title: CloudNet
  contact:
    name: CloudNet
    url: https://cloudnetservice.eu
  license:
    name: Apache license, version 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: '{protocol}://{host}:{port}/api/v3/'
    variables:
      protocol:
        enum:
          - http
          - https
        default: https
      host:
        default: 127.0.0.1
      port:
        default: '2812'
security:
  - bearerAuth: [ ]
...
paths:
...
  /task:
    post:
      requestBody:
        required: true
        description: The json data representing the task to create or update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTask'
      tags:
        - Tasks
      summary: Creates a new task or updates an already existing one
      description: |
        Creates a new task if there is no other task known with the provided name. If there is a task known then the known task is updated using the provided data.

        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:task_write`

          - `cloudnet_rest:task_create`
      responses:
        '204':
          description: The task was created successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    get:
      tags:
        - Tasks
      summary: Lists all tasks that are known by the node handling the request
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:task_read`

          - `cloudnet_rest:task_list`
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceTask'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /task/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: The name of the task to retrieve
        schema:
          type: string
    get:
      tags:
        - Tasks
      summary: Get a service task using the given name
      description: |
        Get a service task using the name given as path parameter. The task name is case-sensitive.

        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:task_read`

          - `cloudnet_rest:task_get`
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTask'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no task with the provided name.
          $ref: '#/components/responses/Problem'
    delete:
      tags:
        - Tasks
      summary: Deletes the task with the provided name
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:task_write`

          - `cloudnet_rest:task_delete`
      responses:
        '204':
          description: The task was deleted successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          description: There is no task with the provided name.
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    Problem:
      description: The action couldn't be exececuted
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Problem:
      description: >-
        An object describing a problem as specific in RFC 9457. Note that this
        object can contain any number of additional properties depending on the
        called route and error cause.
      type: object
      properties:
        type:
          type: string
          format: uri
          description: An absolute URI that identifies the problem type
          default: about:blank
        title:
          type: string
          description: >-
            A short summary of the problem type. Written in English and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized).
          example: Service Unavailable
        status:
          type: integer
          format: int32
          description: >-
            The HTTP status code generated by the origin server for this
            occurrence of the problem.
          minimum: 400
          exclusiveMaximum: 600
          example: 503
        detail:
          type: string
          description: >-
            A human-readable explanation specific to this occurrence of the
            problem
        instance:
          type: string
          format: uri
          description: >-
            An absolute URI reference that identifies the specific occurrence of
            the problem. It may or may not yield further information if
            dereferenced.
...
    JsonDocPropertyable:
      type: object
      properties:
        properties:
          type: object
          examples:
            - { }
    ServiceRemoteInclusion:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            url:
              type: string
              examples:
                - https://my-plugins.de/cool.jar
            destination:
              type: string
              examples:
                - plugins/cool.jar
    ServiceDeployment:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            template:
              $ref: '#/components/schemas/ServiceTemplate'
            excludes:
              type: array
              items:
                type: string
              examples:
                - [ 'spigot.jar', 'plugins/cool.jar' ]
    ServiceConfigurationBase:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            templates:
              type: array
              items:
                $ref: '#/components/schemas/ServiceTemplate'
            deployments:
              type: array
              items:
                $ref: '#/components/schemas/ServiceDeployment'
            includes:
              type: array
              items:
                $ref: '#/components/schemas/ServiceRemoteInclusion'
...
    ServiceTemplate:
      type: object
      properties:
        prefix:
          type: string
          examples:
            - Global
        name:
          type: string
          examples:
            - server
        storage:
          type: string
          examples:
            - local
        priority:
          type: integer
          examples:
            - 5
        alwaysCopyToStaticServices:
          type: boolean
          examples:
            - false
...
    ProcessConfiguration:
      type: object
      properties:
        environment:
          type: string
          examples:
            - MINECRAFT_SERVER
        maxHeapMemorySize:
          type: integer
          examples:
            - 512
        jvmOptions:
          type: array
          items:
            type: string
          examples:
            - [ Option1, Option2 ]
        processParameters:
          type: array
          items:
            type: string
          examples:
            - [ Parameter1, Parameter2 ]
        environmentVariables:
          type: object
          additionalProperties:
            type: string
          examples:
            - MYSQL_PASSWORD: PASSWORD
    ServiceTask:
      type: object
      allOf:
        - $ref: '#/components/schemas/ServiceConfigurationBase'
        - properties:
            name:
              type: string
              examples:
                - Lobby
              pattern: ^[a-zA-Z0-9._\-*]*$
            runtime:
              type: string
              examples:
                - jvm
            javaCommand:
              type: string
              examples:
                - java
            nameSplitter:
              type: string
              examples:
                - '-'
            disableIpRewrite:
              type: boolean
              examples:
                - true
            maintenance:
              type: boolean
              examples:
                - false
            autoDeleteOnStop:
              type: boolean
              examples:
                - false
            staticServices:
              type: boolean
              examples:
                - true
            groups:
              type: array
              items:
                type: string
                examples:
                  - Lobby
            associatedNodes:
              type: array
              items:
                type: string
                examples:
                  - Node-1
            deletedFilesAfterStop:
              type: array
              items:
                type: string
              examples:
                - server.properties
                - plugins/cool/config.json
            processConfiguration:
              $ref: '#/components/schemas/ProcessConfiguration'
            startPort:
              type: integer
              examples:
                - 25565
            minServiceCount:
              type: integer
              examples:
                - 2
...